Use single listener for like/unlike events
